pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "traffic-research"
version = "0.1.0"
description = "Ultralytics-based traffic trajectory research pipeline"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "ultralytics>=8.3.0",
  "numpy>=1.26",
  "pandas>=2.2",
  "scikit-learn>=1.4",
  "scipy>=1.11",
  "hydra-core>=1.3",
  "pyarrow>=15.0",
  "omegaconf>=2.3",
  "rich>=13.7",
  "matplotlib>=3.8",
  "tqdm>=4.66",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "mypy>=1.10",
  "ruff>=0.6",
  "black>=24.8",
  "isort>=5.13",
  "pre-commit>=3.7",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 100

[tool.isort]
profile = "black"

[tool.ruff]
line-length = 100

[tool.mypy]
python_version = "3.11"
warn_unused_ignores = true
strict_optional = true
